requirements.txt
asgiref==3.8.1
dj-database-url==2.1.0
Django==5.0.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
gunicorn==22.0.0
Pillow==10.4.0
psycopg2-binary==2.9.9
pytz==2024.1
sqlparse==0.5.0
whitenoise==6.7.0